+++ module/apmsql/apmsql_test.go
@@ -21,7 +21,7 @@ func TestPingContext(t *testing.T) {
 	defer db.Close()
 
 	db.Ping() // connect
-	tx := withTransaction(t, func(ctx context.Context) {
+	tx, _ := withTransaction(t, func(ctx context.Context) {
 		err := db.PingContext(ctx)
 		assert.NoError(t, err)
 	})
@@ -36,7 +36,7 @@ func TestExecContext(t *testing.T) {
 	defer db.Close()
 
 	db.Ping() // connect
-	tx := withTransaction(t, func(ctx context.Context) {
+	tx, _ := withTransaction(t, func(ctx context.Context) {
 		_, err := db.ExecContext(ctx, "CREATE TABLE foo (bar INT)")
 		require.NoError(t, err)
 	})
@@ -53,7 +53,7 @@ func TestQueryContext(t *testing.T) {
 	_, err = db.Exec("CREATE TABLE foo (bar INT)")
 	require.NoError(t, err)
 
-	tx := withTransaction(t, func(ctx context.Context) {
+	tx, _ := withTransaction(t, func(ctx context.Context) {
 		rows, err := db.QueryContext(ctx, "SELECT * FROM foo")
 		require.NoError(t, err)
 		rows.Close()
@@ -78,7 +78,7 @@ func TestPrepareContext(t *testing.T) {
 	defer db.Close()
 
 	db.Ping() // connect
-	tx := withTransaction(t, func(ctx context.Context) {
+	tx, _ := withTransaction(t, func(ctx context.Context) {
 		stmt, err := db.PrepareContext(ctx, "CREATE TABLE foo (bar INT)")
 		require.NoError(t, err)
 		defer stmt.Close()
@@ -102,7 +102,7 @@ func TestStmtExecContext(t *testing.T) {
 	require.NoError(t, err)
 	defer stmt.Close()
 
-	tx := withTransaction(t, func(ctx context.Context) {
+	tx, _ := withTransaction(t, func(ctx context.Context) {
 		_, err = stmt.ExecContext(ctx, sql.Named("ceil", 999))
 		require.NoError(t, err)
 	})
@@ -123,7 +123,7 @@ func TestStmtQueryContext(t *testing.T) {
 	require.NoError(t, err)
 	defer stmt.Close()
 
-	tx := withTransaction(t, func(ctx context.Context) {
+	tx, _ := withTransaction(t, func(ctx context.Context) {
 		rows, err := stmt.QueryContext(ctx)
 		require.NoError(t, err)
 		rows.Close()
@@ -145,7 +145,7 @@ func TestTxStmtQueryContext(t *testing.T) {
 	require.NoError(t, err)
 	defer stmt.Close()
 
-	tx := withTransaction(t, func(ctx context.Context) {
+	tx, _ := withTransaction(t, func(ctx context.Context) {
 		tx, err := db.BeginTx(ctx, nil)
 		require.NoError(t, err)
 		defer tx.Rollback()
@@ -160,7 +160,24 @@ func TestTxStmtQueryContext(t *testing.T) {
 	assert.Equal(t, "db.sqlite3.query", tx.Spans[0].Type)
 }
 
-func withTransaction(t *testing.T, f func(ctx context.Context)) model.Transaction {
+func TestCaptureErrors(t *testing.T) {
+	db, err := apmsql.Open("sqlite3", ":memory:")
+	require.NoError(t, err)
+	defer db.Close()
+
+	db.Ping() // connect
+	tx, errors := withTransaction(t, func(ctx context.Context) {
+		_, err := db.QueryContext(ctx, "SELECT * FROM thin_air")
+		require.Error(t, err)
+	})
+	require.Len(t, tx.Spans, 1)
+	require.Len(t, errors, 1)
+	assert.Equal(t, "SELECT FROM thin_air", tx.Spans[0].Name)
+	assert.Equal(t, "db.sqlite3.query", tx.Spans[0].Type)
+	assert.Equal(t, "no such table: thin_air", errors[0].Exception.Message)
+}
+
+func withTransaction(t *testing.T, f func(ctx context.Context)) (model.Transaction, []*model.Error) {
 	tracer, transport := transporttest.NewRecorderTracer()
 	defer tracer.Close()
 
@@ -171,8 +188,11 @@ func withTransaction(t *testing.T, f func(ctx context.Context)) model.Transactio
 	tx.End()
 	tracer.Flush(nil)
 	payloads := transport.Payloads()
-	require.Len(t, payloads, 1)
-	transactions := payloads[0].Transactions()
+	var errors []*model.Error
+	if len(payloads) == 2 {
+		errors = payloads[0].Errors()
+	}
+	transactions := payloads[len(payloads)-1].Transactions()
 	require.Len(t, transactions, 1)
-	return transactions[0]
+	return transactions[0], errors
 }